Create standard_batch_trim
This batch file will trim from the front and back of a movie file. Amount to trim from the front is set at line 23 at -ss 00:00:00.0000. This is HH:MM:SS.xxxx
Amount to trim from the back end is set at line 15 at "ws=%%a-00.00", where 00.00 is MM.SS, this can become triple digit to read MMM.SS if need-be.