Create batch_trim_to_time
Similar to the standard batch trim, however this one will trim to a set time in seconds, so you can set the amount to trim from the start at line 15 as usual, and at line 23 you set in seconds where you want the movie to end. This is useful in some scenarios where trimming backwards fails or becomes too bulky.